refactor: move compareDependencyVersions into cli-build
