+++ .changeset/orange-dolls-smell.md
@@ -1,0 +1,6 @@
+---
+"@sanity/cli-build": major
+"@sanity/cli": patch
+---
+
+Move compareDependencyVersions into cli-build package. This change cleans up an unused export getModuleUrl from cli-build.
