+++ packages/@sanity/cli-build/src/util/compareDependencyVersions.ts
@@ -1,14 +1,26 @@
 import path from 'node:path'
 
-import {
-  type CompareDependencyVersions,
-  type CompareDependencyVersionsResult,
-  getModuleUrl,
-  type UnresolvedPrerelease,
-} from '@sanity/cli-build/_internal/build'
 import {getLocalPackageVersion, readPackageJson} from '@sanity/cli-core'
 import {createRequester} from '@sanity/cli-core/request'
 import {coerce, eq, prerelease, parse as semverParse} from 'semver'
+
+import {getModuleUrl} from '../actions/build/getAutoUpdatesImportMap.js'
+
+export interface CompareDependencyVersions {
+  installed: string
+  pkg: string
+  remote: string
+}
+
+export interface UnresolvedPrerelease {
+  pkg: string
+  version: string
+}
+
+export interface CompareDependencyVersionsResult {
+  mismatched: Array<CompareDependencyVersions>
+  unresolvedPrerelease: Array<UnresolvedPrerelease>
+}
 
 const defaultRequester = createRequester({
   middleware: {httpErrors: false, promise: {onlyBody: false}},
